third_party/WebKit/LayoutTests/css-parser/color3_keywords-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
Color test for resources/css-parsing-tests/color3_keywords.json

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colorTest.parseColor("transparent") is "rgba(0, 0, 0, 0)"
PASS colorTest.parseColor("Transparent") is "rgba(0, 0, 0, 0)"
PASS colorTest.parseColor("\\transparent") is "rgba(0, 0, 0, 0)"
PASS colorTest.parseColor("\\74 ransparent") is "rgba(0, 0, 0, 0)"
PASS colorTest.parseColor("ransparent") is "parse error"
PASS colorTest.parseColor("black") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("bLack") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("b\\lack") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("b\\6C ack") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("back") is "parse error"
PASS colorTest.parseColor("blacK") is "parse error"
PASS colorTest.parseColor("silver") is "rgb(192, 192, 192)"
PASS colorTest.parseColor("siLver") is "rgb(192, 192, 192)"
PASS colorTest.parseColor("si\\lver") is "rgb(192, 192, 192)"
PASS colorTest.parseColor("si\\6C ver") is "rgb(192, 192, 192)"
PASS colorTest.parseColor("siver") is "parse error"
PASS colorTest.parseColor("gray") is "rgb(128, 128, 128)"
PASS colorTest.parseColor("graY") is "rgb(128, 128, 128)"
PASS colorTest.parseColor("gra\\y") is "rgb(128, 128, 128)"
PASS colorTest.parseColor("gra\\79 ") is "rgb(128, 128, 128)"
PASS colorTest.parseColor("gra") is "parse error"
PASS colorTest.parseColor("white") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("whitE") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("whit\\65 ") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("whit") is "parse error"
PASS colorTest.parseColor("maroon") is "rgb(128, 0, 0)"
PASS colorTest.parseColor("marooN") is "rgb(128, 0, 0)"
PASS colorTest.parseColor("maroo\\n") is "rgb(128, 0, 0)"
PASS colorTest.parseColor("maroo\\6E ") is "rgb(128, 0, 0)"
PASS colorTest.parseColor("maroo") is "parse error"
PASS colorTest.parseColor("red") is "rgb(255, 0, 0)"
PASS colorTest.parseColor("Red") is "rgb(255, 0, 0)"
PASS colorTest.parseColor("\\red") is "rgb(255, 0, 0)"
PASS colorTest.parseColor("\\72 ed") is "rgb(255, 0, 0)"
PASS colorTest.parseColor("ed") is "parse error"
PASS colorTest.parseColor("purple") is "rgb(128, 0, 128)"
PASS colorTest.parseColor("pUrple") is "rgb(128, 0, 128)"
PASS colorTest.parseColor("p\\urple") is "rgb(128, 0, 128)"
PASS colorTest.parseColor("p\\75 rple") is "rgb(128, 0, 128)"
PASS colorTest.parseColor("prple") is "parse error"
PASS colorTest.parseColor("fuchsia") is "rgb(255, 0, 255)"
PASS colorTest.parseColor("fUchsia") is "rgb(255, 0, 255)"
PASS colorTest.parseColor("f\\uchsia") is "rgb(255, 0, 255)"
PASS colorTest.parseColor("f\\75 chsia") is "rgb(255, 0, 255)"
PASS colorTest.parseColor("fchsia") is "parse error"
PASS colorTest.parseColor("green") is "rgb(0, 128, 0)"
PASS colorTest.parseColor("greeN") is "rgb(0, 128, 0)"
PASS colorTest.parseColor("gree\\n") is "rgb(0, 128, 0)"
PASS colorTest.parseColor("gree\\6E ") is "rgb(0, 128, 0)"
PASS colorTest.parseColor("gree") is "parse error"
PASS colorTest.parseColor("lime") is "rgb(0, 255, 0)"
PASS colorTest.parseColor("liMe") is "rgb(0, 255, 0)"
PASS colorTest.parseColor("li\\me") is "rgb(0, 255, 0)"
PASS colorTest.parseColor("li\\6D e") is "rgb(0, 255, 0)"
PASS colorTest.parseColor("lie") is "parse error"
PASS colorTest.parseColor("olive") is "rgb(128, 128, 0)"
PASS colorTest.parseColor("oLive") is "rgb(128, 128, 0)"
PASS colorTest.parseColor("o\\live") is "rgb(128, 128, 0)"
PASS colorTest.parseColor("o\\6C ive") is "rgb(128, 128, 0)"
PASS colorTest.parseColor("oive") is "parse error"
PASS colorTest.parseColor("yellow") is "rgb(255, 255, 0)"
PASS colorTest.parseColor("Yellow") is "rgb(255, 255, 0)"
PASS colorTest.parseColor("\\yellow") is "rgb(255, 255, 0)"
PASS colorTest.parseColor("\\79 ellow") is "rgb(255, 255, 0)"
PASS colorTest.parseColor("ellow") is "parse error"
PASS colorTest.parseColor("navy") is "rgb(0, 0, 128)"
PASS colorTest.parseColor("nAvy") is "rgb(0, 0, 128)"
PASS colorTest.parseColor("n\\61 vy") is "rgb(0, 0, 128)"
PASS colorTest.parseColor("nvy") is "parse error"
PASS colorTest.parseColor("blue") is "rgb(0, 0, 255)"
PASS colorTest.parseColor("blUe") is "rgb(0, 0, 255)"
PASS colorTest.parseColor("bl\\ue") is "rgb(0, 0, 255)"
PASS colorTest.parseColor("bl\\75 e") is "rgb(0, 0, 255)"
PASS colorTest.parseColor("ble") is "parse error"
PASS colorTest.parseColor("teal") is "rgb(0, 128, 128)"
PASS colorTest.parseColor("teaL") is "rgb(0, 128, 128)"
PASS colorTest.parseColor("tea\\l") is "rgb(0, 128, 128)"
PASS colorTest.parseColor("tea\\6C ") is "rgb(0, 128, 128)"
PASS colorTest.parseColor("tea") is "parse error"
PASS colorTest.parseColor("aqua") is "rgb(0, 255, 255)"
PASS colorTest.parseColor("Aqua") is "rgb(0, 255, 255)"
PASS colorTest.parseColor("\\61 qua") is "rgb(0, 255, 255)"
PASS colorTest.parseColor("qua") is "parse error"
PASS colorTest.parseColor("aliceblue") is "rgb(240, 248, 255)"
PASS colorTest.parseColor("alicebluE") is "rgb(240, 248, 255)"
PASS colorTest.parseColor("aliceblu\\65 ") is "rgb(240, 248, 255)"
PASS colorTest.parseColor("aliceblu") is "parse error"
PASS colorTest.parseColor("antiquewhite") is "rgb(250, 235, 215)"
PASS colorTest.parseColor("antiquEwhite") is "rgb(250, 235, 215)"
PASS colorTest.parseColor("antiqu\\65 white") is "rgb(250, 235, 215)"
PASS colorTest.parseColor("antiquwhite") is "parse error"
PASS colorTest.parseColor("aqua") is "rgb(0, 255, 255)"
PASS colorTest.parseColor("aquA") is "rgb(0, 255, 255)"
PASS colorTest.parseColor("aqu\\61 ") is "rgb(0, 255, 255)"
PASS colorTest.parseColor("aqu") is "parse error"
PASS colorTest.parseColor("aquamarine") is "rgb(127, 255, 212)"
PASS colorTest.parseColor("Aquamarine") is "rgb(127, 255, 212)"
PASS colorTest.parseColor("\\61 quamarine") is "rgb(127, 255, 212)"
PASS colorTest.parseColor("quamarine") is "parse error"
PASS colorTest.parseColor("azure") is "rgb(240, 255, 255)"
PASS colorTest.parseColor("aZure") is "rgb(240, 255, 255)"
PASS colorTest.parseColor("a\\zure") is "rgb(240, 255, 255)"
PASS colorTest.parseColor("a\\7A ure") is "rgb(240, 255, 255)"
PASS colorTest.parseColor("aure") is "parse error"
PASS colorTest.parseColor("beige") is "rgb(245, 245, 220)"
PASS colorTest.parseColor("beIge") is "rgb(245, 245, 220)"
PASS colorTest.parseColor("be\\ige") is "rgb(245, 245, 220)"
PASS colorTest.parseColor("be\\69 ge") is "rgb(245, 245, 220)"
PASS colorTest.parseColor("bege") is "parse error"
PASS colorTest.parseColor("bisque") is "rgb(255, 228, 196)"
PASS colorTest.parseColor("bisquE") is "rgb(255, 228, 196)"
PASS colorTest.parseColor("bisqu\\65 ") is "rgb(255, 228, 196)"
PASS colorTest.parseColor("bisqu") is "parse error"
PASS colorTest.parseColor("black") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("blacK") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("blac\\k") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("blac\\6B ") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("blac") is "parse error"
PASS colorTest.parseColor("blacK") is "parse error"
PASS colorTest.parseColor("blanchedalmond") is "rgb(255, 235, 205)"
PASS colorTest.parseColor("blanchedalmOnd") is "rgb(255, 235, 205)"
PASS colorTest.parseColor("blanchedalm\\ond") is "rgb(255, 235, 205)"
PASS colorTest.parseColor("blanchedalm\\6F nd") is "rgb(255, 235, 205)"
PASS colorTest.parseColor("blanchedalmnd") is "parse error"
PASS colorTest.parseColor("blue") is "rgb(0, 0, 255)"
PASS colorTest.parseColor("blUe") is "rgb(0, 0, 255)"
PASS colorTest.parseColor("bl\\ue") is "rgb(0, 0, 255)"
PASS colorTest.parseColor("bl\\75 e") is "rgb(0, 0, 255)"
PASS colorTest.parseColor("ble") is "parse error"
PASS colorTest.parseColor("blueviolet") is "rgb(138, 43, 226)"
PASS colorTest.parseColor("bluevioLet") is "rgb(138, 43, 226)"
PASS colorTest.parseColor("bluevio\\let") is "rgb(138, 43, 226)"
PASS colorTest.parseColor("bluevio\\6C et") is "rgb(138, 43, 226)"
PASS colorTest.parseColor("bluevioet") is "parse error"
PASS colorTest.parseColor("brown") is "rgb(165, 42, 42)"
PASS colorTest.parseColor("broWn") is "rgb(165, 42, 42)"
PASS colorTest.parseColor("bro\\wn") is "rgb(165, 42, 42)"
PASS colorTest.parseColor("bro\\77 n") is "rgb(165, 42, 42)"
PASS colorTest.parseColor("bron") is "parse error"
PASS colorTest.parseColor("burlywood") is "rgb(222, 184, 135)"
PASS colorTest.parseColor("buRlywood") is "rgb(222, 184, 135)"
PASS colorTest.parseColor("bu\\rlywood") is "rgb(222, 184, 135)"
PASS colorTest.parseColor("bu\\72 lywood") is "rgb(222, 184, 135)"
PASS colorTest.parseColor("bulywood") is "parse error"
PASS colorTest.parseColor("cadetblue") is "rgb(95, 158, 160)"
PASS colorTest.parseColor("cadEtblue") is "rgb(95, 158, 160)"
PASS colorTest.parseColor("cad\\65 tblue") is "rgb(95, 158, 160)"
PASS colorTest.parseColor("cadtblue") is "parse error"
PASS colorTest.parseColor("chartreuse") is "rgb(127, 255, 0)"
PASS colorTest.parseColor("cHartreuse") is "rgb(127, 255, 0)"
PASS colorTest.parseColor("c\\hartreuse") is "rgb(127, 255, 0)"
PASS colorTest.parseColor("c\\68 artreuse") is "rgb(127, 255, 0)"
PASS colorTest.parseColor("cartreuse") is "parse error"
PASS colorTest.parseColor("chocolate") is "rgb(210, 105, 30)"
PASS colorTest.parseColor("chocoLate") is "rgb(210, 105, 30)"
PASS colorTest.parseColor("choco\\late") is "rgb(210, 105, 30)"
PASS colorTest.parseColor("choco\\6C ate") is "rgb(210, 105, 30)"
PASS colorTest.parseColor("chocoate") is "parse error"
PASS colorTest.parseColor("coral") is "rgb(255, 127, 80)"
PASS colorTest.parseColor("corAl") is "rgb(255, 127, 80)"
PASS colorTest.parseColor("cor\\61 l") is "rgb(255, 127, 80)"
PASS colorTest.parseColor("corl") is "parse error"
PASS colorTest.parseColor("cornflowerblue") is "rgb(100, 149, 237)"
PASS colorTest.parseColor("cornflOwerblue") is "rgb(100, 149, 237)"
PASS colorTest.parseColor("cornfl\\owerblue") is "rgb(100, 149, 237)"
PASS colorTest.parseColor("cornfl\\6F werblue") is "rgb(100, 149, 237)"
PASS colorTest.parseColor("cornflwerblue") is "parse error"
PASS colorTest.parseColor("cornsilk") is "rgb(255, 248, 220)"
PASS colorTest.parseColor("corNsilk") is "rgb(255, 248, 220)"
PASS colorTest.parseColor("cor\\nsilk") is "rgb(255, 248, 220)"
PASS colorTest.parseColor("cor\\6E silk") is "rgb(255, 248, 220)"
PASS colorTest.parseColor("corsilk") is "parse error"
PASS colorTest.parseColor("cornsilK") is "parse error"
PASS colorTest.parseColor("crimson") is "rgb(220, 20, 60)"
PASS colorTest.parseColor("cRimson") is "rgb(220, 20, 60)"
PASS colorTest.parseColor("c\\rimson") is "rgb(220, 20, 60)"
PASS colorTest.parseColor("c\\72 imson") is "rgb(220, 20, 60)"
PASS colorTest.parseColor("cimson") is "parse error"
PASS colorTest.parseColor("cyan") is "rgb(0, 255, 255)"
PASS colorTest.parseColor("cYan") is "rgb(0, 255, 255)"
PASS colorTest.parseColor("c\\yan") is "rgb(0, 255, 255)"
PASS colorTest.parseColor("c\\79 an") is "rgb(0, 255, 255)"
PASS colorTest.parseColor("can") is "parse error"
PASS colorTest.parseColor("darkblue") is "rgb(0, 0, 139)"
PASS colorTest.parseColor("darkblUe") is "rgb(0, 0, 139)"
PASS colorTest.parseColor("darkbl\\ue") is "rgb(0, 0, 139)"
PASS colorTest.parseColor("darkbl\\75 e") is "rgb(0, 0, 139)"
PASS colorTest.parseColor("darkble") is "parse error"
PASS colorTest.parseColor("darKblue") is "parse error"
PASS colorTest.parseColor("darkcyan") is "rgb(0, 139, 139)"
PASS colorTest.parseColor("darkcyaN") is "rgb(0, 139, 139)"
PASS colorTest.parseColor("darkcya\\n") is "rgb(0, 139, 139)"
PASS colorTest.parseColor("darkcya\\6E ") is "rgb(0, 139, 139)"
PASS colorTest.parseColor("darkcya") is "parse error"
PASS colorTest.parseColor("darKcyan") is "parse error"
PASS colorTest.parseColor("darkgoldenrod") is "rgb(184, 134, 11)"
PASS colorTest.parseColor("dArkgoldenrod") is "rgb(184, 134, 11)"
PASS colorTest.parseColor("d\\61 rkgoldenrod") is "rgb(184, 134, 11)"
PASS colorTest.parseColor("drkgoldenrod") is "parse error"
PASS colorTest.parseColor("darKgoldenrod") is "parse error"
PASS colorTest.parseColor("darkgray") is "rgb(169, 169, 169)"
PASS colorTest.parseColor("dArkgray") is "rgb(169, 169, 169)"
PASS colorTest.parseColor("d\\61 rkgray") is "rgb(169, 169, 169)"
PASS colorTest.parseColor("drkgray") is "parse error"
PASS colorTest.parseColor("darKgray") is "parse error"
PASS colorTest.parseColor("darkgreen") is "rgb(0, 100, 0)"
PASS colorTest.parseColor("darkgrEen") is "rgb(0, 100, 0)"
PASS colorTest.parseColor("darkgr\\65 en") is "rgb(0, 100, 0)"
PASS colorTest.parseColor("darkgren") is "parse error"
PASS colorTest.parseColor("darKgreen") is "parse error"
PASS colorTest.parseColor("darkgrey") is "rgb(169, 169, 169)"
PASS colorTest.parseColor("darKgrey") is "rgb(169, 169, 169)"
PASS colorTest.parseColor("dar\\kgrey") is "rgb(169, 169, 169)"
PASS colorTest.parseColor("dar\\6B grey") is "rgb(169, 169, 169)"
PASS colorTest.parseColor("dargrey") is "parse error"
PASS colorTest.parseColor("darKgrey") is "parse error"
PASS colorTest.parseColor("darkkhaki") is "rgb(189, 183, 107)"
PASS colorTest.parseColor("darkkhakI") is "rgb(189, 183, 107)"
PASS colorTest.parseColor("darkkhak\\i") is "rgb(189, 183, 107)"
PASS colorTest.parseColor("darkkhak\\69 ") is "rgb(189, 183, 107)"
PASS colorTest.parseColor("darkkhak") is "parse error"
PASS colorTest.parseColor("darKKhaKi") is "parse error"
PASS colorTest.parseColor("darkmagenta") is "rgb(139, 0, 139)"
PASS colorTest.parseColor("dArkmagenta") is "rgb(139, 0, 139)"
PASS colorTest.parseColor("d\\61 rkmagenta") is "rgb(139, 0, 139)"
PASS colorTest.parseColor("drkmagenta") is "parse error"
PASS colorTest.parseColor("darKmagenta") is "parse error"
PASS colorTest.parseColor("darkolivegreen") is "rgb(85, 107, 47)"
PASS colorTest.parseColor("darkOlivegreen") is "rgb(85, 107, 47)"
PASS colorTest.parseColor("dark\\olivegreen") is "rgb(85, 107, 47)"
PASS colorTest.parseColor("dark\\6F livegreen") is "rgb(85, 107, 47)"
PASS colorTest.parseColor("darklivegreen") is "parse error"
PASS colorTest.parseColor("darKolivegreen") is "parse error"
PASS colorTest.parseColor("darkorange") is "rgb(255, 140, 0)"
PASS colorTest.parseColor("darkoraNge") is "rgb(255, 140, 0)"
PASS colorTest.parseColor("darkora\\nge") is "rgb(255, 140, 0)"
PASS colorTest.parseColor("darkora\\6E ge") is "rgb(255, 140, 0)"
PASS colorTest.parseColor("darkorage") is "parse error"
PASS colorTest.parseColor("darKorange") is "parse error"
PASS colorTest.parseColor("darkorchid") is "rgb(153, 50, 204)"
PASS colorTest.parseColor("darkorchId") is "rgb(153, 50, 204)"
PASS colorTest.parseColor("darkorch\\id") is "rgb(153, 50, 204)"
PASS colorTest.parseColor("darkorch\\69 d") is "rgb(153, 50, 204)"
PASS colorTest.parseColor("darkorchd") is "parse error"
PASS colorTest.parseColor("darKorchid") is "parse error"
PASS colorTest.parseColor("darkred") is "rgb(139, 0, 0)"
PASS colorTest.parseColor("Darkred") is "rgb(139, 0, 0)"
PASS colorTest.parseColor("\\64 arkred") is "rgb(139, 0, 0)"
PASS colorTest.parseColor("arkred") is "parse error"
PASS colorTest.parseColor("darKred") is "parse error"
PASS colorTest.parseColor("darksalmon") is "rgb(233, 150, 122)"
PASS colorTest.parseColor("Darksalmon") is "rgb(233, 150, 122)"
PASS colorTest.parseColor("\\64 arksalmon") is "rgb(233, 150, 122)"
PASS colorTest.parseColor("arksalmon") is "parse error"
PASS colorTest.parseColor("darKsalmon") is "parse error"
PASS colorTest.parseColor("darkseagreen") is "rgb(143, 188, 143)"
PASS colorTest.parseColor("darKseagreen") is "rgb(143, 188, 143)"
PASS colorTest.parseColor("dar\\kseagreen") is "rgb(143, 188, 143)"
PASS colorTest.parseColor("dar\\6B seagreen") is "rgb(143, 188, 143)"
PASS colorTest.parseColor("darseagreen") is "parse error"
PASS colorTest.parseColor("darKseagreen") is "parse error"
PASS colorTest.parseColor("darkslateblue") is "rgb(72, 61, 139)"
PASS colorTest.parseColor("Darkslateblue") is "rgb(72, 61, 139)"
PASS colorTest.parseColor("\\64 arkslateblue") is "rgb(72, 61, 139)"
PASS colorTest.parseColor("arkslateblue") is "parse error"
PASS colorTest.parseColor("darKslateblue") is "parse error"
PASS colorTest.parseColor("darkslategray") is "rgb(47, 79, 79)"
PASS colorTest.parseColor("dArkslategray") is "rgb(47, 79, 79)"
PASS colorTest.parseColor("d\\61 rkslategray") is "rgb(47, 79, 79)"
PASS colorTest.parseColor("drkslategray") is "parse error"
PASS colorTest.parseColor("darKslategray") is "parse error"
PASS colorTest.parseColor("darkslategrey") is "rgb(47, 79, 79)"
PASS colorTest.parseColor("daRkslategrey") is "rgb(47, 79, 79)"
PASS colorTest.parseColor("da\\rkslategrey") is "rgb(47, 79, 79)"
PASS colorTest.parseColor("da\\72 kslategrey") is "rgb(47, 79, 79)"
PASS colorTest.parseColor("dakslategrey") is "parse error"
PASS colorTest.parseColor("darKslategrey") is "parse error"
PASS colorTest.parseColor("darkturquoise") is "rgb(0, 206, 209)"
PASS colorTest.parseColor("darKturquoise") is "rgb(0, 206, 209)"
PASS colorTest.parseColor("dar\\kturquoise") is "rgb(0, 206, 209)"
PASS colorTest.parseColor("dar\\6B turquoise") is "rgb(0, 206, 209)"
PASS colorTest.parseColor("darturquoise") is "parse error"
PASS colorTest.parseColor("darKturquoise") is "parse error"
PASS colorTest.parseColor("darkviolet") is "rgb(148, 0, 211)"
PASS colorTest.parseColor("darkviOlet") is "rgb(148, 0, 211)"
PASS colorTest.parseColor("darkvi\\olet") is "rgb(148, 0, 211)"
PASS colorTest.parseColor("darkvi\\6F let") is "rgb(148, 0, 211)"
PASS colorTest.parseColor("darkvilet") is "parse error"
PASS colorTest.parseColor("darKviolet") is "parse error"
PASS colorTest.parseColor("deeppink") is "rgb(255, 20, 147)"
PASS colorTest.parseColor("dEeppink") is "rgb(255, 20, 147)"
PASS colorTest.parseColor("d\\65 eppink") is "rgb(255, 20, 147)"
PASS colorTest.parseColor("deppink") is "parse error"
PASS colorTest.parseColor("deeppinK") is "parse error"
PASS colorTest.parseColor("deepskyblue") is "rgb(0, 191, 255)"
PASS colorTest.parseColor("deePskyblue") is "rgb(0, 191, 255)"
PASS colorTest.parseColor("dee\\pskyblue") is "rgb(0, 191, 255)"
PASS colorTest.parseColor("dee\\70 skyblue") is "rgb(0, 191, 255)"
PASS colorTest.parseColor("deeskyblue") is "parse error"
PASS colorTest.parseColor("deepsKyblue") is "parse error"
PASS colorTest.parseColor("dimgray") is "rgb(105, 105, 105)"
PASS colorTest.parseColor("dimGray") is "rgb(105, 105, 105)"
PASS colorTest.parseColor("dim\\gray") is "rgb(105, 105, 105)"
PASS colorTest.parseColor("dim\\67 ray") is "rgb(105, 105, 105)"
PASS colorTest.parseColor("dimray") is "parse error"
PASS colorTest.parseColor("dimgrey") is "rgb(105, 105, 105)"
PASS colorTest.parseColor("dimgRey") is "rgb(105, 105, 105)"
PASS colorTest.parseColor("dimg\\rey") is "rgb(105, 105, 105)"
PASS colorTest.parseColor("dimg\\72 ey") is "rgb(105, 105, 105)"
PASS colorTest.parseColor("dimgey") is "parse error"
PASS colorTest.parseColor("dodgerblue") is "rgb(30, 144, 255)"
PASS colorTest.parseColor("dOdgerblue") is "rgb(30, 144, 255)"
PASS colorTest.parseColor("d\\odgerblue") is "rgb(30, 144, 255)"
PASS colorTest.parseColor("d\\6F dgerblue") is "rgb(30, 144, 255)"
PASS colorTest.parseColor("ddgerblue") is "parse error"
PASS colorTest.parseColor("firebrick") is "rgb(178, 34, 34)"
PASS colorTest.parseColor("firebricK") is "rgb(178, 34, 34)"
PASS colorTest.parseColor("firebric\\k") is "rgb(178, 34, 34)"
PASS colorTest.parseColor("firebric\\6B ") is "rgb(178, 34, 34)"
PASS colorTest.parseColor("firebric") is "parse error"
PASS colorTest.parseColor("firebricK") is "parse error"
PASS colorTest.parseColor("floralwhite") is "rgb(255, 250, 240)"
PASS colorTest.parseColor("floralwhIte") is "rgb(255, 250, 240)"
PASS colorTest.parseColor("floralwh\\ite") is "rgb(255, 250, 240)"
PASS colorTest.parseColor("floralwh\\69 te") is "rgb(255, 250, 240)"
PASS colorTest.parseColor("floralwhte") is "parse error"
PASS colorTest.parseColor("forestgreen") is "rgb(34, 139, 34)"
PASS colorTest.parseColor("forestgreEn") is "rgb(34, 139, 34)"
PASS colorTest.parseColor("forestgre\\65 n") is "rgb(34, 139, 34)"
PASS colorTest.parseColor("forestgren") is "parse error"
PASS colorTest.parseColor("fuchsia") is "rgb(255, 0, 255)"
PASS colorTest.parseColor("fuChsia") is "rgb(255, 0, 255)"
PASS colorTest.parseColor("fu\\63 hsia") is "rgb(255, 0, 255)"
PASS colorTest.parseColor("fuhsia") is "parse error"
PASS colorTest.parseColor("gainsboro") is "rgb(220, 220, 220)"
PASS colorTest.parseColor("gaiNsboro") is "rgb(220, 220, 220)"
PASS colorTest.parseColor("gai\\nsboro") is "rgb(220, 220, 220)"
PASS colorTest.parseColor("gai\\6E sboro") is "rgb(220, 220, 220)"
PASS colorTest.parseColor("gaisboro") is "parse error"
PASS colorTest.parseColor("ghostwhite") is "rgb(248, 248, 255)"
PASS colorTest.parseColor("ghostwhIte") is "rgb(248, 248, 255)"
PASS colorTest.parseColor("ghostwh\\ite") is "rgb(248, 248, 255)"
PASS colorTest.parseColor("ghostwh\\69 te") is "rgb(248, 248, 255)"
PASS colorTest.parseColor("ghostwhte") is "parse error"
PASS colorTest.parseColor("gold") is "rgb(255, 215, 0)"
PASS colorTest.parseColor("Gold") is "rgb(255, 215, 0)"
PASS colorTest.parseColor("\\gold") is "rgb(255, 215, 0)"
PASS colorTest.parseColor("\\67 old") is "rgb(255, 215, 0)"
PASS colorTest.parseColor("old") is "parse error"
PASS colorTest.parseColor("goldenrod") is "rgb(218, 165, 32)"
PASS colorTest.parseColor("goldenRod") is "rgb(218, 165, 32)"
PASS colorTest.parseColor("golden\\rod") is "rgb(218, 165, 32)"
PASS colorTest.parseColor("golden\\72 od") is "rgb(218, 165, 32)"
PASS colorTest.parseColor("goldenod") is "parse error"
PASS colorTest.parseColor("gray") is "rgb(128, 128, 128)"
PASS colorTest.parseColor("grAy") is "rgb(128, 128, 128)"
PASS colorTest.parseColor("gr\\61 y") is "rgb(128, 128, 128)"
PASS colorTest.parseColor("gry") is "parse error"
PASS colorTest.parseColor("green") is "rgb(0, 128, 0)"
PASS colorTest.parseColor("gReen") is "rgb(0, 128, 0)"
PASS colorTest.parseColor("g\\reen") is "rgb(0, 128, 0)"
PASS colorTest.parseColor("g\\72 een") is "rgb(0, 128, 0)"
PASS colorTest.parseColor("geen") is "parse error"
PASS colorTest.parseColor("greenyellow") is "rgb(173, 255, 47)"
PASS colorTest.parseColor("greenyEllow") is "rgb(173, 255, 47)"
PASS colorTest.parseColor("greeny\\65 llow") is "rgb(173, 255, 47)"
PASS colorTest.parseColor("greenyllow") is "parse error"
PASS colorTest.parseColor("grey") is "rgb(128, 128, 128)"
PASS colorTest.parseColor("gRey") is "rgb(128, 128, 128)"
PASS colorTest.parseColor("g\\rey") is "rgb(128, 128, 128)"
PASS colorTest.parseColor("g\\72 ey") is "rgb(128, 128, 128)"
PASS colorTest.parseColor("gey") is "parse error"
PASS colorTest.parseColor("honeydew") is "rgb(240, 255, 240)"
PASS colorTest.parseColor("hoNeydew") is "rgb(240, 255, 240)"
PASS colorTest.parseColor("ho\\neydew") is "rgb(240, 255, 240)"
PASS colorTest.parseColor("ho\\6E eydew") is "rgb(240, 255, 240)"
PASS colorTest.parseColor("hoeydew") is "parse error"
PASS colorTest.parseColor("hotpink") is "rgb(255, 105, 180)"
PASS colorTest.parseColor("hotpiNk") is "rgb(255, 105, 180)"
PASS colorTest.parseColor("hotpi\\nk") is "rgb(255, 105, 180)"
PASS colorTest.parseColor("hotpi\\6E k") is "rgb(255, 105, 180)"
PASS colorTest.parseColor("hotpik") is "parse error"
PASS colorTest.parseColor("hotpinK") is "parse error"
PASS colorTest.parseColor("indianred") is "rgb(205, 92, 92)"
PASS colorTest.parseColor("indiAnred") is "rgb(205, 92, 92)"
PASS colorTest.parseColor("indi\\61 nred") is "rgb(205, 92, 92)"
PASS colorTest.parseColor("indinred") is "parse error"
PASS colorTest.parseColor("indigo") is "rgb(75, 0, 130)"
PASS colorTest.parseColor("indigO") is "rgb(75, 0, 130)"
PASS colorTest.parseColor("indig\\o") is "rgb(75, 0, 130)"
PASS colorTest.parseColor("indig\\6F ") is "rgb(75, 0, 130)"
PASS colorTest.parseColor("indig") is "parse error"
PASS colorTest.parseColor("ivory") is "rgb(255, 255, 240)"
PASS colorTest.parseColor("ivoRy") is "rgb(255, 255, 240)"
PASS colorTest.parseColor("ivo\\ry") is "rgb(255, 255, 240)"
PASS colorTest.parseColor("ivo\\72 y") is "rgb(255, 255, 240)"
PASS colorTest.parseColor("ivoy") is "parse error"
PASS colorTest.parseColor("khaki") is "rgb(240, 230, 140)"
PASS colorTest.parseColor("khakI") is "rgb(240, 230, 140)"
PASS colorTest.parseColor("khak\\i") is "rgb(240, 230, 140)"
PASS colorTest.parseColor("khak\\69 ") is "rgb(240, 230, 140)"
PASS colorTest.parseColor("khak") is "parse error"
PASS colorTest.parseColor("KhaKi") is "parse error"
PASS colorTest.parseColor("lavender") is "rgb(230, 230, 250)"
PASS colorTest.parseColor("Lavender") is "rgb(230, 230, 250)"
PASS colorTest.parseColor("\\lavender") is "rgb(230, 230, 250)"
PASS colorTest.parseColor("\\6C avender") is "rgb(230, 230, 250)"
PASS colorTest.parseColor("avender") is "parse error"
PASS colorTest.parseColor("lavenderblush") is "rgb(255, 240, 245)"
PASS colorTest.parseColor("lavEnderblush") is "rgb(255, 240, 245)"
PASS colorTest.parseColor("lav\\65 nderblush") is "rgb(255, 240, 245)"
PASS colorTest.parseColor("lavnderblush") is "parse error"
PASS colorTest.parseColor("lawngreen") is "rgb(124, 252, 0)"
PASS colorTest.parseColor("lAwngreen") is "rgb(124, 252, 0)"
PASS colorTest.parseColor("l\\61 wngreen") is "rgb(124, 252, 0)"
PASS colorTest.parseColor("lwngreen") is "parse error"
PASS colorTest.parseColor("lemonchiffon") is "rgb(255, 250, 205)"
PASS colorTest.parseColor("lemonchiffoN") is "rgb(255, 250, 205)"
PASS colorTest.parseColor("lemonchiffo\\n") is "rgb(255, 250, 205)"
PASS colorTest.parseColor("lemonchiffo\\6E ") is "rgb(255, 250, 205)"
PASS colorTest.parseColor("lemonchiffo") is "parse error"
PASS colorTest.parseColor("lightblue") is "rgb(173, 216, 230)"
PASS colorTest.parseColor("ligHtblue") is "rgb(173, 216, 230)"
PASS colorTest.parseColor("lig\\htblue") is "rgb(173, 216, 230)"
PASS colorTest.parseColor("lig\\68 tblue") is "rgb(173, 216, 230)"
PASS colorTest.parseColor("ligtblue") is "parse error"
PASS colorTest.parseColor("lightcoral") is "rgb(240, 128, 128)"
PASS colorTest.parseColor("lightCoral") is "rgb(240, 128, 128)"
PASS colorTest.parseColor("light\\63 oral") is "rgb(240, 128, 128)"
PASS colorTest.parseColor("lightoral") is "parse error"
PASS colorTest.parseColor("lightcyan") is "rgb(224, 255, 255)"
PASS colorTest.parseColor("lightCyan") is "rgb(224, 255, 255)"
PASS colorTest.parseColor("light\\63 yan") is "rgb(224, 255, 255)"
PASS colorTest.parseColor("lightyan") is "parse error"
PASS colorTest.parseColor("lightgoldenrodyellow") is "rgb(250, 250, 210)"
PASS colorTest.parseColor("lightgoLdenrodyellow") is "rgb(250, 250, 210)"
PASS colorTest.parseColor("lightgo\\ldenrodyellow") is "rgb(250, 250, 210)"
PASS colorTest.parseColor("lightgo\\6C denrodyellow") is "rgb(250, 250, 210)"
PASS colorTest.parseColor("lightgodenrodyellow") is "parse error"
PASS colorTest.parseColor("lightgray") is "rgb(211, 211, 211)"
PASS colorTest.parseColor("lightgrAy") is "rgb(211, 211, 211)"
PASS colorTest.parseColor("lightgr\\61 y") is "rgb(211, 211, 211)"
PASS colorTest.parseColor("lightgry") is "parse error"
PASS colorTest.parseColor("lightgreen") is "rgb(144, 238, 144)"
PASS colorTest.parseColor("lightgreeN") is "rgb(144, 238, 144)"
PASS colorTest.parseColor("lightgree\\n") is "rgb(144, 238, 144)"
PASS colorTest.parseColor("lightgree\\6E ") is "rgb(144, 238, 144)"
PASS colorTest.parseColor("lightgree") is "parse error"
PASS colorTest.parseColor("lightgrey") is "rgb(211, 211, 211)"
PASS colorTest.parseColor("Lightgrey") is "rgb(211, 211, 211)"
PASS colorTest.parseColor("\\lightgrey") is "rgb(211, 211, 211)"
PASS colorTest.parseColor("\\6C ightgrey") is "rgb(211, 211, 211)"
PASS colorTest.parseColor("ightgrey") is "parse error"
PASS colorTest.parseColor("lightpink") is "rgb(255, 182, 193)"
PASS colorTest.parseColor("lIghtpink") is "rgb(255, 182, 193)"
PASS colorTest.parseColor("l\\ightpink") is "rgb(255, 182, 193)"
PASS colorTest.parseColor("l\\69 ghtpink") is "rgb(255, 182, 193)"
PASS colorTest.parseColor("lghtpink") is "parse error"
PASS colorTest.parseColor("lightpinK") is "parse error"
PASS colorTest.parseColor("lightsalmon") is "rgb(255, 160, 122)"
PASS colorTest.parseColor("lighTsalmon") is "rgb(255, 160, 122)"
PASS colorTest.parseColor("ligh\\tsalmon") is "rgb(255, 160, 122)"
PASS colorTest.parseColor("ligh\\74 salmon") is "rgb(255, 160, 122)"
PASS colorTest.parseColor("lighsalmon") is "parse error"
PASS colorTest.parseColor("lightseagreen") is "rgb(32, 178, 170)"
PASS colorTest.parseColor("liGhtseagreen") is "rgb(32, 178, 170)"
PASS colorTest.parseColor("li\\ghtseagreen") is "rgb(32, 178, 170)"
PASS colorTest.parseColor("li\\67 htseagreen") is "rgb(32, 178, 170)"
PASS colorTest.parseColor("lihtseagreen") is "parse error"
PASS colorTest.parseColor("lightskyblue") is "rgb(135, 206, 250)"
PASS colorTest.parseColor("lightskyblUe") is "rgb(135, 206, 250)"
PASS colorTest.parseColor("lightskybl\\ue") is "rgb(135, 206, 250)"
PASS colorTest.parseColor("lightskybl\\75 e") is "rgb(135, 206, 250)"
PASS colorTest.parseColor("lightskyble") is "parse error"
PASS colorTest.parseColor("lightsKyblue") is "parse error"
PASS colorTest.parseColor("lightslategray") is "rgb(119, 136, 153)"
PASS colorTest.parseColor("lightslategRay") is "rgb(119, 136, 153)"
PASS colorTest.parseColor("lightslateg\\ray") is "rgb(119, 136, 153)"
PASS colorTest.parseColor("lightslateg\\72 ay") is "rgb(119, 136, 153)"
PASS colorTest.parseColor("lightslategay") is "parse error"
PASS colorTest.parseColor("lightslategrey") is "rgb(119, 136, 153)"
PASS colorTest.parseColor("lightslategrEy") is "rgb(119, 136, 153)"
PASS colorTest.parseColor("lightslategr\\65 y") is "rgb(119, 136, 153)"
PASS colorTest.parseColor("lightslategry") is "parse error"
PASS colorTest.parseColor("lightsteelblue") is "rgb(176, 196, 222)"
PASS colorTest.parseColor("lightsteelbluE") is "rgb(176, 196, 222)"
PASS colorTest.parseColor("lightsteelblu\\65 ") is "rgb(176, 196, 222)"
PASS colorTest.parseColor("lightsteelblu") is "parse error"
PASS colorTest.parseColor("lightyellow") is "rgb(255, 255, 224)"
PASS colorTest.parseColor("lightyelloW") is "rgb(255, 255, 224)"
PASS colorTest.parseColor("lightyello\\w") is "rgb(255, 255, 224)"
PASS colorTest.parseColor("lightyello\\77 ") is "rgb(255, 255, 224)"
PASS colorTest.parseColor("lightyello") is "parse error"
PASS colorTest.parseColor("lime") is "rgb(0, 255, 0)"
PASS colorTest.parseColor("limE") is "rgb(0, 255, 0)"
PASS colorTest.parseColor("lim\\65 ") is "rgb(0, 255, 0)"
PASS colorTest.parseColor("lim") is "parse error"
PASS colorTest.parseColor("limegreen") is "rgb(50, 205, 50)"
PASS colorTest.parseColor("lImegreen") is "rgb(50, 205, 50)"
PASS colorTest.parseColor("l\\imegreen") is "rgb(50, 205, 50)"
PASS colorTest.parseColor("l\\69 megreen") is "rgb(50, 205, 50)"
PASS colorTest.parseColor("lmegreen") is "parse error"
PASS colorTest.parseColor("linen") is "rgb(250, 240, 230)"
PASS colorTest.parseColor("lInen") is "rgb(250, 240, 230)"
PASS colorTest.parseColor("l\\inen") is "rgb(250, 240, 230)"
PASS colorTest.parseColor("l\\69 nen") is "rgb(250, 240, 230)"
PASS colorTest.parseColor("lnen") is "parse error"
PASS colorTest.parseColor("magenta") is "rgb(255, 0, 255)"
PASS colorTest.parseColor("mageNta") is "rgb(255, 0, 255)"
PASS colorTest.parseColor("mage\\nta") is "rgb(255, 0, 255)"
PASS colorTest.parseColor("mage\\6E ta") is "rgb(255, 0, 255)"
PASS colorTest.parseColor("mageta") is "parse error"
PASS colorTest.parseColor("maroon") is "rgb(128, 0, 0)"
PASS colorTest.parseColor("mAroon") is "rgb(128, 0, 0)"
PASS colorTest.parseColor("m\\61 roon") is "rgb(128, 0, 0)"
PASS colorTest.parseColor("mroon") is "parse error"
PASS colorTest.parseColor("mediumaquamarine") is "rgb(102, 205, 170)"
PASS colorTest.parseColor("mediumaqUamarine") is "rgb(102, 205, 170)"
PASS colorTest.parseColor("mediumaq\\uamarine") is "rgb(102, 205, 170)"
PASS colorTest.parseColor("mediumaq\\75 amarine") is "rgb(102, 205, 170)"
PASS colorTest.parseColor("mediumaqamarine") is "parse error"
PASS colorTest.parseColor("mediumblue") is "rgb(0, 0, 205)"
PASS colorTest.parseColor("mediuMblue") is "rgb(0, 0, 205)"
PASS colorTest.parseColor("mediu\\mblue") is "rgb(0, 0, 205)"
PASS colorTest.parseColor("mediu\\6D blue") is "rgb(0, 0, 205)"
PASS colorTest.parseColor("mediublue") is "parse error"
PASS colorTest.parseColor("mediumorchid") is "rgb(186, 85, 211)"
PASS colorTest.parseColor("mediumorchId") is "rgb(186, 85, 211)"
PASS colorTest.parseColor("mediumorch\\id") is "rgb(186, 85, 211)"
PASS colorTest.parseColor("mediumorch\\69 d") is "rgb(186, 85, 211)"
PASS colorTest.parseColor("mediumorchd") is "parse error"
PASS colorTest.parseColor("mediumpurple") is "rgb(147, 112, 219)"
PASS colorTest.parseColor("mediumpurplE") is "rgb(147, 112, 219)"
PASS colorTest.parseColor("mediumpurpl\\65 ") is "rgb(147, 112, 219)"
PASS colorTest.parseColor("mediumpurpl") is "parse error"
PASS colorTest.parseColor("mediumseagreen") is "rgb(60, 179, 113)"
PASS colorTest.parseColor("mediumseagReen") is "rgb(60, 179, 113)"
PASS colorTest.parseColor("mediumseag\\reen") is "rgb(60, 179, 113)"
PASS colorTest.parseColor("mediumseag\\72 een") is "rgb(60, 179, 113)"
PASS colorTest.parseColor("mediumseageen") is "parse error"
PASS colorTest.parseColor("mediumslateblue") is "rgb(123, 104, 238)"
PASS colorTest.parseColor("mediUmslateblue") is "rgb(123, 104, 238)"
PASS colorTest.parseColor("medi\\umslateblue") is "rgb(123, 104, 238)"
PASS colorTest.parseColor("medi\\75 mslateblue") is "rgb(123, 104, 238)"
PASS colorTest.parseColor("medimslateblue") is "parse error"
PASS colorTest.parseColor("mediumspringgreen") is "rgb(0, 250, 154)"
PASS colorTest.parseColor("mediumspRinggreen") is "rgb(0, 250, 154)"
PASS colorTest.parseColor("mediumsp\\ringgreen") is "rgb(0, 250, 154)"
PASS colorTest.parseColor("mediumsp\\72 inggreen") is "rgb(0, 250, 154)"
PASS colorTest.parseColor("mediumspinggreen") is "parse error"
PASS colorTest.parseColor("mediumturquoise") is "rgb(72, 209, 204)"
PASS colorTest.parseColor("mediumTurquoise") is "rgb(72, 209, 204)"
PASS colorTest.parseColor("medium\\turquoise") is "rgb(72, 209, 204)"
PASS colorTest.parseColor("medium\\74 urquoise") is "rgb(72, 209, 204)"
PASS colorTest.parseColor("mediumurquoise") is "parse error"
PASS colorTest.parseColor("mediumvioletred") is "rgb(199, 21, 133)"
PASS colorTest.parseColor("mediumvIoletred") is "rgb(199, 21, 133)"
PASS colorTest.parseColor("mediumv\\ioletred") is "rgb(199, 21, 133)"
PASS colorTest.parseColor("mediumv\\69 oletred") is "rgb(199, 21, 133)"
PASS colorTest.parseColor("mediumvoletred") is "parse error"
PASS colorTest.parseColor("midnightblue") is "rgb(25, 25, 112)"
PASS colorTest.parseColor("midniGhtblue") is "rgb(25, 25, 112)"
PASS colorTest.parseColor("midni\\ghtblue") is "rgb(25, 25, 112)"
PASS colorTest.parseColor("midni\\67 htblue") is "rgb(25, 25, 112)"
PASS colorTest.parseColor("midnihtblue") is "parse error"
PASS colorTest.parseColor("mintcream") is "rgb(245, 255, 250)"
PASS colorTest.parseColor("mintcrEam") is "rgb(245, 255, 250)"
PASS colorTest.parseColor("mintcr\\65 am") is "rgb(245, 255, 250)"
PASS colorTest.parseColor("mintcram") is "parse error"
PASS colorTest.parseColor("mistyrose") is "rgb(255, 228, 225)"
PASS colorTest.parseColor("mistyroSe") is "rgb(255, 228, 225)"
PASS colorTest.parseColor("mistyro\\se") is "rgb(255, 228, 225)"
PASS colorTest.parseColor("mistyro\\73 e") is "rgb(255, 228, 225)"
PASS colorTest.parseColor("mistyroe") is "parse error"
PASS colorTest.parseColor("moccasin") is "rgb(255, 228, 181)"
PASS colorTest.parseColor("moccAsin") is "rgb(255, 228, 181)"
PASS colorTest.parseColor("mocc\\61 sin") is "rgb(255, 228, 181)"
PASS colorTest.parseColor("moccsin") is "parse error"
PASS colorTest.parseColor("navajowhite") is "rgb(255, 222, 173)"
PASS colorTest.parseColor("navajowHite") is "rgb(255, 222, 173)"
PASS colorTest.parseColor("navajow\\hite") is "rgb(255, 222, 173)"
PASS colorTest.parseColor("navajow\\68 ite") is "rgb(255, 222, 173)"
PASS colorTest.parseColor("navajowite") is "parse error"
PASS colorTest.parseColor("navy") is "rgb(0, 0, 128)"
PASS colorTest.parseColor("naVy") is "rgb(0, 0, 128)"
PASS colorTest.parseColor("na\\vy") is "rgb(0, 0, 128)"
PASS colorTest.parseColor("na\\76 y") is "rgb(0, 0, 128)"
PASS colorTest.parseColor("nay") is "parse error"
PASS colorTest.parseColor("oldlace") is "rgb(253, 245, 230)"
PASS colorTest.parseColor("Oldlace") is "rgb(253, 245, 230)"
PASS colorTest.parseColor("\\oldlace") is "rgb(253, 245, 230)"
PASS colorTest.parseColor("\\6F ldlace") is "rgb(253, 245, 230)"
PASS colorTest.parseColor("ldlace") is "parse error"
PASS colorTest.parseColor("olive") is "rgb(128, 128, 0)"
PASS colorTest.parseColor("Olive") is "rgb(128, 128, 0)"
PASS colorTest.parseColor("\\olive") is "rgb(128, 128, 0)"
PASS colorTest.parseColor("\\6F live") is "rgb(128, 128, 0)"
PASS colorTest.parseColor("live") is "parse error"
PASS colorTest.parseColor("olivedrab") is "rgb(107, 142, 35)"
PASS colorTest.parseColor("olivEdrab") is "rgb(107, 142, 35)"
PASS colorTest.parseColor("oliv\\65 drab") is "rgb(107, 142, 35)"
PASS colorTest.parseColor("olivdrab") is "parse error"
PASS colorTest.parseColor("orange") is "rgb(255, 165, 0)"
PASS colorTest.parseColor("orAnge") is "rgb(255, 165, 0)"
PASS colorTest.parseColor("or\\61 nge") is "rgb(255, 165, 0)"
PASS colorTest.parseColor("ornge") is "parse error"
PASS colorTest.parseColor("orangered") is "rgb(255, 69, 0)"
PASS colorTest.parseColor("orangeRed") is "rgb(255, 69, 0)"
PASS colorTest.parseColor("orange\\red") is "rgb(255, 69, 0)"
PASS colorTest.parseColor("orange\\72 ed") is "rgb(255, 69, 0)"
PASS colorTest.parseColor("orangeed") is "parse error"
PASS colorTest.parseColor("orchid") is "rgb(218, 112, 214)"
PASS colorTest.parseColor("orchId") is "rgb(218, 112, 214)"
PASS colorTest.parseColor("orch\\id") is "rgb(218, 112, 214)"
PASS colorTest.parseColor("orch\\69 d") is "rgb(218, 112, 214)"
PASS colorTest.parseColor("orchd") is "parse error"
PASS colorTest.parseColor("palegoldenrod") is "rgb(238, 232, 170)"
PASS colorTest.parseColor("palegoldEnrod") is "rgb(238, 232, 170)"
PASS colorTest.parseColor("palegold\\65 nrod") is "rgb(238, 232, 170)"
PASS colorTest.parseColor("palegoldnrod") is "parse error"
PASS colorTest.parseColor("palegreen") is "rgb(152, 251, 152)"
PASS colorTest.parseColor("Palegreen") is "rgb(152, 251, 152)"
PASS colorTest.parseColor("\\palegreen") is "rgb(152, 251, 152)"
PASS colorTest.parseColor("\\70 alegreen") is "rgb(152, 251, 152)"
PASS colorTest.parseColor("alegreen") is "parse error"
PASS colorTest.parseColor("paleturquoise") is "rgb(175, 238, 238)"
PASS colorTest.parseColor("paleturquoIse") is "rgb(175, 238, 238)"
PASS colorTest.parseColor("paleturquo\\ise") is "rgb(175, 238, 238)"
PASS colorTest.parseColor("paleturquo\\69 se") is "rgb(175, 238, 238)"
PASS colorTest.parseColor("paleturquose") is "parse error"
PASS colorTest.parseColor("palevioletred") is "rgb(219, 112, 147)"
PASS colorTest.parseColor("palevioletrEd") is "rgb(219, 112, 147)"
PASS colorTest.parseColor("palevioletr\\65 d") is "rgb(219, 112, 147)"
PASS colorTest.parseColor("palevioletrd") is "parse error"
PASS colorTest.parseColor("papayawhip") is "rgb(255, 239, 213)"
PASS colorTest.parseColor("papayawhiP") is "rgb(255, 239, 213)"
PASS colorTest.parseColor("papayawhi\\p") is "rgb(255, 239, 213)"
PASS colorTest.parseColor("papayawhi\\70 ") is "rgb(255, 239, 213)"
PASS colorTest.parseColor("papayawhi") is "parse error"
PASS colorTest.parseColor("peachpuff") is "rgb(255, 218, 185)"
PASS colorTest.parseColor("peacHpuff") is "rgb(255, 218, 185)"
PASS colorTest.parseColor("peac\\hpuff") is "rgb(255, 218, 185)"
PASS colorTest.parseColor("peac\\68 puff") is "rgb(255, 218, 185)"
PASS colorTest.parseColor("peacpuff") is "parse error"
PASS colorTest.parseColor("peru") is "rgb(205, 133, 63)"
PASS colorTest.parseColor("perU") is "rgb(205, 133, 63)"
PASS colorTest.parseColor("per\\u") is "rgb(205, 133, 63)"
PASS colorTest.parseColor("per\\75 ") is "rgb(205, 133, 63)"
PASS colorTest.parseColor("per") is "parse error"
PASS colorTest.parseColor("pink") is "rgb(255, 192, 203)"
PASS colorTest.parseColor("Pink") is "rgb(255, 192, 203)"
PASS colorTest.parseColor("\\pink") is "rgb(255, 192, 203)"
PASS colorTest.parseColor("\\70 ink") is "rgb(255, 192, 203)"
PASS colorTest.parseColor("ink") is "parse error"
PASS colorTest.parseColor("pinK") is "parse error"
PASS colorTest.parseColor("plum") is "rgb(221, 160, 221)"
PASS colorTest.parseColor("pLum") is "rgb(221, 160, 221)"
PASS colorTest.parseColor("p\\lum") is "rgb(221, 160, 221)"
PASS colorTest.parseColor("p\\6C um") is "rgb(221, 160, 221)"
PASS colorTest.parseColor("pum") is "parse error"
PASS colorTest.parseColor("powderblue") is "rgb(176, 224, 230)"
PASS colorTest.parseColor("powdErblue") is "rgb(176, 224, 230)"
PASS colorTest.parseColor("powd\\65 rblue") is "rgb(176, 224, 230)"
PASS colorTest.parseColor("powdrblue") is "parse error"
PASS colorTest.parseColor("purple") is "rgb(128, 0, 128)"
PASS colorTest.parseColor("purPle") is "rgb(128, 0, 128)"
PASS colorTest.parseColor("pur\\ple") is "rgb(128, 0, 128)"
PASS colorTest.parseColor("pur\\70 le") is "rgb(128, 0, 128)"
PASS colorTest.parseColor("purle") is "parse error"
PASS colorTest.parseColor("red") is "rgb(255, 0, 0)"
PASS colorTest.parseColor("rEd") is "rgb(255, 0, 0)"
PASS colorTest.parseColor("r\\65 d") is "rgb(255, 0, 0)"
PASS colorTest.parseColor("rd") is "parse error"
PASS colorTest.parseColor("rosybrown") is "rgb(188, 143, 143)"
PASS colorTest.parseColor("roSybrown") is "rgb(188, 143, 143)"
PASS colorTest.parseColor("ro\\sybrown") is "rgb(188, 143, 143)"
PASS colorTest.parseColor("ro\\73 ybrown") is "rgb(188, 143, 143)"
PASS colorTest.parseColor("roybrown") is "parse error"
PASS colorTest.parseColor("royalblue") is "rgb(65, 105, 225)"
PASS colorTest.parseColor("royAlblue") is "rgb(65, 105, 225)"
PASS colorTest.parseColor("roy\\61 lblue") is "rgb(65, 105, 225)"
PASS colorTest.parseColor("roylblue") is "parse error"
PASS colorTest.parseColor("saddlebrown") is "rgb(139, 69, 19)"
PASS colorTest.parseColor("saddlebRown") is "rgb(139, 69, 19)"
PASS colorTest.parseColor("saddleb\\rown") is "rgb(139, 69, 19)"
PASS colorTest.parseColor("saddleb\\72 own") is "rgb(139, 69, 19)"
PASS colorTest.parseColor("saddlebown") is "parse error"
PASS colorTest.parseColor("salmon") is "rgb(250, 128, 114)"
PASS colorTest.parseColor("saLmon") is "rgb(250, 128, 114)"
PASS colorTest.parseColor("sa\\lmon") is "rgb(250, 128, 114)"
PASS colorTest.parseColor("sa\\6C mon") is "rgb(250, 128, 114)"
PASS colorTest.parseColor("samon") is "parse error"
PASS colorTest.parseColor("sandybrown") is "rgb(244, 164, 96)"
PASS colorTest.parseColor("sAndybrown") is "rgb(244, 164, 96)"
PASS colorTest.parseColor("s\\61 ndybrown") is "rgb(244, 164, 96)"
PASS colorTest.parseColor("sndybrown") is "parse error"
PASS colorTest.parseColor("seagreen") is "rgb(46, 139, 87)"
PASS colorTest.parseColor("seagreEn") is "rgb(46, 139, 87)"
PASS colorTest.parseColor("seagre\\65 n") is "rgb(46, 139, 87)"
PASS colorTest.parseColor("seagren") is "parse error"
PASS colorTest.parseColor("seashell") is "rgb(255, 245, 238)"
PASS colorTest.parseColor("seashelL") is "rgb(255, 245, 238)"
PASS colorTest.parseColor("seashel\\l") is "rgb(255, 245, 238)"
PASS colorTest.parseColor("seashel\\6C ") is "rgb(255, 245, 238)"
PASS colorTest.parseColor("seashel") is "parse error"
PASS colorTest.parseColor("sienna") is "rgb(160, 82, 45)"
PASS colorTest.parseColor("Sienna") is "rgb(160, 82, 45)"
PASS colorTest.parseColor("\\sienna") is "rgb(160, 82, 45)"
PASS colorTest.parseColor("\\73 ienna") is "rgb(160, 82, 45)"
PASS colorTest.parseColor("ienna") is "parse error"
PASS colorTest.parseColor("silver") is "rgb(192, 192, 192)"
PASS colorTest.parseColor("sIlver") is "rgb(192, 192, 192)"
PASS colorTest.parseColor("s\\ilver") is "rgb(192, 192, 192)"
PASS colorTest.parseColor("s\\69 lver") is "rgb(192, 192, 192)"
PASS colorTest.parseColor("slver") is "parse error"
PASS colorTest.parseColor("skyblue") is "rgb(135, 206, 235)"
PASS colorTest.parseColor("skybluE") is "rgb(135, 206, 235)"
PASS colorTest.parseColor("skyblu\\65 ") is "rgb(135, 206, 235)"
PASS colorTest.parseColor("skyblu") is "parse error"
PASS colorTest.parseColor("sKyblue") is "parse error"
PASS colorTest.parseColor("slateblue") is "rgb(106, 90, 205)"
PASS colorTest.parseColor("slaTeblue") is "rgb(106, 90, 205)"
PASS colorTest.parseColor("sla\\teblue") is "rgb(106, 90, 205)"
PASS colorTest.parseColor("sla\\74 eblue") is "rgb(106, 90, 205)"
PASS colorTest.parseColor("slaeblue") is "parse error"
PASS colorTest.parseColor("slategray") is "rgb(112, 128, 144)"
PASS colorTest.parseColor("slatEgray") is "rgb(112, 128, 144)"
PASS colorTest.parseColor("slat\\65 gray") is "rgb(112, 128, 144)"
PASS colorTest.parseColor("slatgray") is "parse error"
PASS colorTest.parseColor("slategrey") is "rgb(112, 128, 144)"
PASS colorTest.parseColor("slateGrey") is "rgb(112, 128, 144)"
PASS colorTest.parseColor("slate\\grey") is "rgb(112, 128, 144)"
PASS colorTest.parseColor("slate\\67 rey") is "rgb(112, 128, 144)"
PASS colorTest.parseColor("slaterey") is "parse error"
PASS colorTest.parseColor("snow") is "rgb(255, 250, 250)"
PASS colorTest.parseColor("snOw") is "rgb(255, 250, 250)"
PASS colorTest.parseColor("sn\\ow") is "rgb(255, 250, 250)"
PASS colorTest.parseColor("sn\\6F w") is "rgb(255, 250, 250)"
PASS colorTest.parseColor("snw") is "parse error"
PASS colorTest.parseColor("springgreen") is "rgb(0, 255, 127)"
PASS colorTest.parseColor("springgrEen") is "rgb(0, 255, 127)"
PASS colorTest.parseColor("springgr\\65 en") is "rgb(0, 255, 127)"
PASS colorTest.parseColor("springgren") is "parse error"
PASS colorTest.parseColor("steelblue") is "rgb(70, 130, 180)"
PASS colorTest.parseColor("steelbluE") is "rgb(70, 130, 180)"
PASS colorTest.parseColor("steelblu\\65 ") is "rgb(70, 130, 180)"
PASS colorTest.parseColor("steelblu") is "parse error"
PASS colorTest.parseColor("tan") is "rgb(210, 180, 140)"
PASS colorTest.parseColor("Tan") is "rgb(210, 180, 140)"
PASS colorTest.parseColor("\\tan") is "rgb(210, 180, 140)"
PASS colorTest.parseColor("\\74 an") is "rgb(210, 180, 140)"
PASS colorTest.parseColor("an") is "parse error"
PASS colorTest.parseColor("teal") is "rgb(0, 128, 128)"
PASS colorTest.parseColor("teAl") is "rgb(0, 128, 128)"
PASS colorTest.parseColor("te\\61 l") is "rgb(0, 128, 128)"
PASS colorTest.parseColor("tel") is "parse error"
PASS colorTest.parseColor("thistle") is "rgb(216, 191, 216)"
PASS colorTest.parseColor("tHistle") is "rgb(216, 191, 216)"
PASS colorTest.parseColor("t\\histle") is "rgb(216, 191, 216)"
PASS colorTest.parseColor("t\\68 istle") is "rgb(216, 191, 216)"
PASS colorTest.parseColor("tistle") is "parse error"
PASS colorTest.parseColor("tomato") is "rgb(255, 99, 71)"
PASS colorTest.parseColor("Tomato") is "rgb(255, 99, 71)"
PASS colorTest.parseColor("\\tomato") is "rgb(255, 99, 71)"
PASS colorTest.parseColor("\\74 omato") is "rgb(255, 99, 71)"
PASS colorTest.parseColor("omato") is "parse error"
PASS colorTest.parseColor("turquoise") is "rgb(64, 224, 208)"
PASS colorTest.parseColor("turqUoise") is "rgb(64, 224, 208)"
PASS colorTest.parseColor("turq\\uoise") is "rgb(64, 224, 208)"
PASS colorTest.parseColor("turq\\75 oise") is "rgb(64, 224, 208)"
PASS colorTest.parseColor("turqoise") is "parse error"
PASS colorTest.parseColor("violet") is "rgb(238, 130, 238)"
PASS colorTest.parseColor("viOlet") is "rgb(238, 130, 238)"
PASS colorTest.parseColor("vi\\olet") is "rgb(238, 130, 238)"
PASS colorTest.parseColor("vi\\6F let") is "rgb(238, 130, 238)"
PASS colorTest.parseColor("vilet") is "parse error"
PASS colorTest.parseColor("wheat") is "rgb(245, 222, 179)"
PASS colorTest.parseColor("wheaT") is "rgb(245, 222, 179)"
PASS colorTest.parseColor("whea\\t") is "rgb(245, 222, 179)"
PASS colorTest.parseColor("whea\\74 ") is "rgb(245, 222, 179)"
PASS colorTest.parseColor("whea") is "parse error"
PASS colorTest.parseColor("white") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("White") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("\\white") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("\\77 hite") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hite") is "parse error"
PASS colorTest.parseColor("whitesmoke") is "rgb(245, 245, 245)"
PASS colorTest.parseColor("wHitesmoke") is "rgb(245, 245, 245)"
PASS colorTest.parseColor("w\\hitesmoke") is "rgb(245, 245, 245)"
PASS colorTest.parseColor("w\\68 itesmoke") is "rgb(245, 245, 245)"
PASS colorTest.parseColor("witesmoke") is "parse error"
PASS colorTest.parseColor("whitesmoKe") is "parse error"
PASS colorTest.parseColor("yellow") is "rgb(255, 255, 0)"
PASS colorTest.parseColor("Yellow") is "rgb(255, 255, 0)"
PASS colorTest.parseColor("\\yellow") is "rgb(255, 255, 0)"
PASS colorTest.parseColor("\\79 ellow") is "rgb(255, 255, 0)"
PASS colorTest.parseColor("ellow") is "parse error"
PASS colorTest.parseColor("yellowgreen") is "rgb(154, 205, 50)"
PASS colorTest.parseColor("yellowgreEn") is "rgb(154, 205, 50)"
PASS colorTest.parseColor("yellowgre\\65 n") is "rgb(154, 205, 50)"
PASS colorTest.parseColor("yellowgren") is "parse error"
PASS successfullyParsed is true

TEST COMPLETE
